pyproject.toml
[project]
name = "fcfmramos"
dependencies = [
    "flask",
    "Flask-WTF",
    "pyrebase4",
    "firebase-admin"
]
version = "0.0.1"
description = "Opiniones Ramos FCFM"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"